+++ Code/Data Classes/DialogueLineData.cs
@@ -1,0 +1,13 @@
+﻿using DiskCardGame;
+
+namespace JLPlugin.Data
+{
+    public class DialogueLineData
+    {
+        public string text;
+        public string emotion;
+        public int speakerIndex;
+        public string storyCondition;
+        public bool storyConditionMustBeMet;
+    }
+}