+++ Code/Utils/JLUtils.cs
@@ -37,6 +37,30 @@ namespace JLPlugin.Utils
             }
         }
 
+        public static void LoadRegionsFromFiles()
+        {
+            foreach ( string file in Directory.EnumerateFiles( Paths.PluginPath, "*_region.jldr", SearchOption.AllDirectories ) )
+            {
+                string filename = file.Substring( file.LastIndexOf( Path.DirectorySeparatorChar ) + 1 );
+
+                string text = File.ReadAllText (file );
+
+                if ( !text.ToLowerInvariant().StartsWith( "#ignore" ) && !text.ToLowerInvariant().StartsWith( "# ignore" ) )
+                {
+                    Data.RegionData region = CreateRegionFromJSON( text );
+
+                    if (region is null)
+                    {
+                        Plugin.Log.LogWarning( $"Failed to load region { filename }" );
+
+                        continue;
+                    }
+
+                    region.GenerateNew();
+                }
+            }
+        }
+
         public static void LoadEncountersFromFiles()
         {
             foreach ( string file in Directory.EnumerateFiles( Paths.PluginPath, "*_encounter.jldr", SearchOption.AllDirectories ) )
@@ -61,6 +85,9 @@ namespace JLPlugin.Utils
 
         public static EncounterData CreateEncounterFromJSON( string jsonString )
             => JSONParser.FromJson<EncounterData>( jsonString );
+
+        public static RegionData CreateRegionFromJSON( string jsonString )
+            => JSONParser.FromJson<RegionData>( jsonString );
 
         public static Texture2D LoadTexture2D( string image )
         {